Arel::Nodes::NotBetween
