+++ lib/arel/extensions/nodes.rb
@@ -4,9 +4,6 @@
 
 module Arel
   module Nodes
-    class NotBetween < Arel::Nodes::Between
-    end
-
     # postgres only: https://www.postgresql.org/docs/9.1/functions-comparison.html
     class BetweenSymmetric < Arel::Nodes::Between
     end
